+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+- ✨(aliases) import existing aliases from dimail
+- ✨(aliases) can create, list and delete aliases #974
+
 ## [1.20.0] - 2025-10-22
 
-- 🐛(mailbox) fix case-sensitive duplicates on display names
+- ✨(models) impose uniqueness on display name, to match ox's constraint
+- 🐛(dimail) catch duplicate displayname error
 - ✨(mailbox) synchronize password of newly created mailbox with Dimail's
 
 ## [1.19.1] - 2025-09-19